+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
-## [Version 1.0.0](https://github.com/dataiku/dss-plugin-sendmail/releases/tag/v1.0.0) - Feature release - 2024-01
+## [Version 1.0.0](https://github.com/dataiku/dss-plugin-sendmail/releases/tag/v1.0.0) - Feature release - 2024-02
 
-- Added templating capabilities, send as HTML
+- REMOVED SUPPORT FOR PYTHON 2.7 AND 3.5. The code env used must be in range python 3.6 - 3.11 inclusive
+
+- Features added
+  - templating capabilities - use JINJA style templating in the email body and subject line
+  - send as HTML
+  - ability to use DSS channels to send emails
 
 ## [Version 0.2.0](https://github.com/dataiku/dss-plugin-sendmail/releases/tag/v0.2.0) - Feature release - 2023-08